refactor: cambio para funcionar appRouter
