+++ src/app/page.module.css
@@ -72,6 +72,7 @@
     font-size: 0.9rem;
     line-height: 1.5;
     max-width: 30ch;
+    text-wrap: balance;
 }
 
 .center {
